refactor: extract pure playback core crate
