+++ crates/yututui-core/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "yututui-core"
+version = "1.6.32"
+edition = "2024"
+publish = false
+description = "Pure playback policies shared by yututui playback owners."
+license = "MIT"
+
+[features]
+default = []
+ts-export = ["dep:ts-rs"]
+
+[dependencies]
+serde = { version = "1.0.228", features = ["derive"] }
+ts-rs = { version = "10", optional = true, features = ["no-serde-warnings"] }
+
+[lints.rust]
+unsafe_op_in_unsafe_fn = "deny"
+
+[lints.clippy]
+undocumented_unsafe_blocks = "deny"
